requirements.txt
# web
uvicorn
fastapi

# database
sqlalchemy
alembic

# core proxmox & ssh
proxmoxer
paramiko

# config
pydantic
pydantic-settings
python-dotenv

# logging
loguru

# testing
httpie